update to win 11 #1
"SUPPORT FOR WINDOWS 11
The VDOT tool now supports Windows 11.
Windows 11 in some respects, reports the same as Windows 10, and currently (as of 5/5/22) has 'ReleaseID' value of '2009'. Until the 'ReleaseID' changes, all new optimizations are going to be included in the 'Configuration Files' folder underneath the '2009' folder. Therefore, the 2009 folder configuration files apply to Windows 10, as well as Windows 11, as of 5/5/22." 
https://github.com/The-Virtual-Desktop-Team/Virtual-Desktop-Optimization-Tool